requirements.txt
# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1

# PDF Extraction
PyMuPDF==1.24.14

# ML / Inference
# ⚠️ Install torch FIRST from CUDA wheel index:
# pip install torch==2.5.1 torchvision==0.20.1 torchaudio==2.5.1 --index-url https://download.pytorch.org/whl/cu121
transformers>=4.47.0
accelerate>=1.2.0
bitsandbytes>=0.45.0
huggingface_hub>=0.26.0

# Utilities
regex==2024.9.11